requirements.txt
tflite_runtime
sounddevice
python-daemon